Run integration tests with mocha.js.
Problem: unclear test execution results.
Cause: just using "assert" is not enough.
Measure: applying `mocha` as a framework for integration tests.

`Mocha` allows to setup a server (sandbox) which makes it suitable to integration tests.
It allows to run everything with a single command `npm test`.
Alternatives: `Jest` is harder to use in the current conditions.
