+++ package.json
@@ -4,10 +4,15 @@
   "scripts": {
     "build": "rustup target add wasm32-unknown-unknown && cargo build --target wasm32-unknown-unknown --release",
     "sandbox": "rm -rf /tmp/near-sandbox && near-sandbox --home /tmp/near-sandbox init && near-sandbox --home /tmp/near-sandbox run",
-    "test": "node tests/sandbox.js"
+    "test": "mocha --require tests/sandbox-setup.js tests"
   },
   "dependencies": {
     "bn.js": "^5.2.0",
     "near-api-js": "^0.44.2"
+  },
+  "devDependencies": {
+    "mocha": "^9.1.3",
+    "near-sandbox": "^0.0.8",
+    "tree-kill": "^1.2.2"
   }
 }
